+++ packages/utilities/src/guards/Rate Limiter/index.ts
@@ -1,3 +1,5 @@
 export * from "./RateLimit.js";
+export * from "./TimeOutEntry.js";
+export * from "./types.js";
 export * from "./utils/ITimedSet.js";
 export * from "./utils/TimedSet.js";
