+++ packages/utilities/src/guards/Rate Limiter/types.ts
@@ -1,0 +1,6 @@
+export enum TIME_UNIT {
+  days,
+  hours,
+  minutes,
+  seconds,
+}
